Switch from make to ninja and stop using shallow checkout
Shallow checkouts do not behave well when trying to work with branches.  And "get fetch --unshallow" doesn't appear to actually work.

None of LLVM's builtbots exercise make any more. ninja is faster with lots of cores (*much* faster). And the log output is a heck of lot easier to follow.